Got the MySQL Cluster install working
modified the bootstrap file to get the install procedure working. Makes
use of gdebi as Cluster is not included in either the Ubuntu/Debian
repos or the MySQL ones.

This means that on init, the bootstrap has to download a 500mb .deb
file, which looks to have pretty much the entire MySQL suite.

It also installs itself under /opt :-(